users authorization logic OK

+++ controllers/users.js
@@ -3,6 +3,8 @@ const dotenv = require('dotenv');
 const sendNotification = require('../tools/ntfy');
 const User = require('../models/User');
 const storeController = require('./stores');
+const { v4: uuidv4 } = require('uuid');
+
 
 
 dotenv.config();
@@ -17,6 +19,9 @@ const getUser = async (req, res, next) => {
     const id = req.params.id;
 
     const user = await collection.findOne({ _id: new ObjectId(id) });
+    if (req.session.user.role === 'manager' && user.store_id !== req.session.user.store_id) {
+      return res.status(403).json({ message: 'Forbidden' });
+    }
 
     if (!user) {
       return res.status(404).json({ message: 'User not found' });
@@ -61,6 +66,15 @@ const updateUser = async (req, res, next) => {
       }
     }
 
+  
+    const userToUpdate = await collection.findOne({ _id: new ObjectId(id) });
+    if(req.session.user.role === 'customer' && userToUpdate._id !== req.session.user._id) {
+      return res.status(403).json({ message: 'Forbidden' });
+    }
+    if (req.session.user.role === 'manager' && userToUpdate.store_id !== req.session.user.store_id) {
+      return res.status(403).json({ message: 'Forbidden' });
+    }
+
     const updatedUser = collection.findOneAndUpdate({ _id: new ObjectId(id) }, { $set: user });
 
     if (!updatedUser) {
@@ -80,6 +94,14 @@ const createUser = async (req, res, next) => {
   const collection = await database.collection('users');
   let user = req.body;
   
+  if(req.session.user !== undefined) {
+    if(req.session.user.role !== 'admin') {
+      if(req.session.user.store_id !== user.store_id) {
+        return res.status(403).json({ message: 'Forbidden' });
+      }
+    }
+  }
+
   try {
     if (!user.github_id) {
       if (user.store_id) {
@@ -87,19 +109,28 @@ const createUser = async (req, res, next) => {
 
         const store = await storeController.getStore(input, res, next);
         if (!store) {
-          return;  
+          return res.status(400).json({ message: 'Store not found' });  
         }
       }
 
       const newUser = new User(user);
       await newUser.validate();
     }
-
+    
     if (!user.role) {
       user.role = 'customer';
     }
 
+    if(user.api_key === undefined) {
+      user.api_key = uuidv4();
+    }
+
     user.active = true;
+
+    const existingUser = await collection.findOne({ email: user.email });
+    if (existingUser) {
+      return res.status(400).json({ message: 'Email already exists' });
+    }
 
     const response = await collection.insertOne(user);
     const output = response.insertedId.toString();
@@ -129,6 +160,15 @@ const deleteUser = async (req, res, next) => {
   const collection = await database.collection('users')
   try {
     const id = req.params.id;
+
+    const userToDelete = await collection.findOne({ _id: new ObjectId(id) });
+
+    if(req.session.user.role === 'customer' && userToDelete._id !== req.session.user._id) {
+      return res.status(403).json({ message: 'Forbidden' });
+    }
+    if (req.session.user.role === 'manager' && userToDelete.store_id !== req.session.user.store_id) {
+      return res.status(403).json({ message: 'Forbidden' });
+    }
 
     const deletedUser = collection.findOneAndDelete({ _id: new ObjectId(id) });
 
@@ -160,7 +200,55 @@ const getUserByGithubId = async (id) => {
   }
 };
 
-// TODO: Get user by api key
+const getUserByIdAndApiKey = async (id, apiKey) => {
+  const database = await mongodb.getDb();
+  const collection = await database.collection('users')
+  try {
+
+    const user = await collection.findOne({ _id: new ObjectId(id), api_key: apiKey });
+
+    if (!user) {
+      return null;
+    }
+    return user;
+  } catch (err) {
+    throw new Error(err.message);
+  }
+};
+
+const getUsersByStoreId = async (id) => {
+  const database = await mongodb.getDb();
+  const collection = await database.collection('users')
+  try {
+
+    const users = await collection.find({ store_id: id }).toArray();
+
+    if (!users) {
+      return null;
+    }
+    return users;
+  } catch (err) {
+    throw new Error(err.message);
+  }
+};
+
+const getUserByEmailAndPassword = async (email, password) => {
+  const database = await mongodb.getDb();
+  const collection = await database.collection('users')
+  try {
+
+    
+    const user  = await collection.findOne({ email: email, password: password });
+    
+    if (!user) {
+      return null;
+    }
+    return user;
+  }
+  catch (err) {
+    throw new Error(err.message);
+  }
+}
 
 module.exports = {
   getUser,
@@ -168,5 +256,8 @@ module.exports = {
   updateUser,
   createUser,
   deleteUser,
-  getUserByGithubId
+  getUserByGithubId,
+  getUserByIdAndApiKey,
+  getUsersByStoreId,
+  getUserByEmailAndPassword
 };
